Started Working on the Libary Feature
Initated work on the Music Libary and File Saving and publishing. This will work as a client server based system where the client authenticates with the server in order to acces the entire libary os BoBd Music avalaible

The Libary will be useing TLS and RSA encryption to encrypt all communications between client and server. And passwords will be stored by encrypting them by using a algorythm called salting. 

Passwords are inaccesable without the password itself to decrypt it.